refactor: Input 소비처를 디자인시스템 컴포넌트로 치환
