+++ src/pages/addAccountStep/AddAccountStepPage.tsx
@@ -3,9 +3,8 @@ import { useState } from 'react';
 import Header from '@/shared/ui/Header';
 import { ArrowDown, ArrowLeft } from '@/shared/assets/svgs/icon';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
-import { Button } from '@/shared/design-system/ui';
+import { Button, Input } from '@/shared/design-system/ui';
 import DescriptionField from '@/shared/ui/DescriptionField';
-import Input from '@/shared/ui/Input';
 import { BoundaryError } from '@/shared/types/error.type';
 import usePutUpdateAccount from '@/features/expense-management/api/usePutUpdateAccount';
 import useDisclosure from './hooks/useDisclosure';
@@ -70,7 +69,7 @@ function AddAccountStepPage({ onNext, onBack }: AddAccountStepProps) {
           placeholder="은행 선택"
           value={bankName}
           onClick={handleBankInputClick}
-          icon={<ArrowDown width={24} />}
+          trailingIcon={<ArrowDown width={24} />}
           readOnly
         />
         <BankNameDrawer
