+++ src/pages/memberSetup/ui/AddMember/index.tsx
@@ -2,11 +2,10 @@ import { useForm } from 'react-hook-form';
 import * as z from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { Member } from '@/entities/member/model/member.type';
+import { Button, Input } from '@/shared/design-system/ui';
 import Text from '@/shared/ui/Text';
 import Profile from '@/shared/ui/Profile';
 import InputGroup from '@/shared/ui/InputGroup';
-import Input from '@/shared/ui/Input';
-import { Button } from '@/shared/design-system/ui';
 import Flex from '@/shared/ui/Flex';
 import { showToast } from '@/shared/ui/Toast';
 import useAddGroupMember from './api/useAddGroupMember';
