+++ src/pages/passwordSetup/PasswordSetupPage.tsx
@@ -4,10 +4,9 @@ import { z } from 'zod';
 import { zodResolver } from '@hookform/resolvers/zod';
 import { useNavigate } from 'react-router';
 import { ArrowLeft } from '@/shared/assets/svgs/icon';
+import { Button, Input } from '@/shared/design-system/ui';
 import Header from '@/shared/ui/Header';
 import DescriptionField from '@/shared/ui/DescriptionField';
-import Input from '@/shared/ui/Input';
-import { Button } from '@/shared/design-system/ui';
 import { BottomButtonContainer } from '@/shared/styles/bottomButton.styles';
 import Text from '@/shared/ui/Text';
 import * as S from './PasswordSetupPage.styles';
